requirements.txt
discord.py==2.4.0
paramiko==3.4.0
asyncpg==0.29.0